Re-write some trip-planner functionalities in order to work when some content is dynamically generated

+++ js/utilities/utils.js
@@ -1,6 +1,8 @@
-export function clearLocalStorage() {
-    for(let i = 0; i < localStorage.length; i++) {
-        localStorage.clear();
-        window.location.reload();
+export function clearLocalStorage(e) {
+    if(e.target.classList.contains('trip-planner__add-subdestination-input')) {
+        for(let i = 0; i < localStorage.length; i++) {
+            localStorage.clear();
+            window.location.reload();
+        }
     }
 }
